Split the auth api logic into its own file
